+++ lib/src/url.dart
@@ -1,5 +1,0 @@
-// Home page url.
-final rootUrl = 'https://www.vicentecaycedo.com';
-
-// Path to apps json.
-final appJsonPath = '/apps.json';