+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.1]
+
+* **Breaking change:** Removed app icon image getter.
+* Fixed: Using full url for app icons since root url is not public.
+
 ## [1.0.0]
 
 * Initial Release.